requirements.txt
# Monitoring
snakeviz

# Config system
hydra-core

# Code quality
black

# Logging
wandb
tensorboardX
tensorboard
gputil

# Utils
matplotlib 
tqdm
imageio[ffmpeg]

# Common libraries
numpy

# Project libraries
# torch
transformers
openai
huggingface-hub
gymnasium
pygame
minigrid